requirements.txt
matplotlib==3.8.2
numpy==1.24.2
opencv_python==4.7.0.68
opencv_python_headless==4.9.0.80
Pillow==9.4.0
Pillow==11.0.0
PyQt5==5.15.11
PyQt5_sip==12.11.0
